+++ Touhou Spell Up!!!/Assets/Scripts/DanmakuGame/Player/Player.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+[CreateAssetMenu(fileName = "Player", menuName = "Touhou-Spell-Up/Player", order = 1)]
+public class Player : ScriptableObject
+{
+    [SerializeField] private PlayerProperty playerProperty;
+    public PlayerProperty PlayerProperty => playerProperty;
+}